Move `test_merge_async_iterators`
